perbaikan migrasi

+++ database/migrations/2024_10_19_164548_create_operasional_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->date('tanggal');
             $table->string('keterangan');
             $table->string('kategori');
-            $table->integer('jumlah');
+            $table->decimal('jumlah', 10,2);
             $table->timestamps();
         });
     }
